Add files via upload
Door sensor connected to a Raspberry pi that triggers several events, each one as a separate process as not to hold up new door switch activity.  Once door is opened a file is processed and commands are sent to other Raspberry pis with cameras.  The remote Rpis are listening on a particular port.  Once activity happens the camera will take a picture.  The remote Rpis are also listening on another port.  Once activity is heard on that port it will upload the files to dropbox using Dropbox_Uploader.sh